# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Session cookie (browser) or Bearer token (scripts, mobile). Cookie-authenticated POST/PUT/PATCH/DELETE must send X-Rigel-Client.",
        "title": "RigelLedger API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign in",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.loginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "429": {
                        "description": "too many failures; see Retry-After",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
...
        "/api/books/{bookID}/accounts/{accountID}/cost": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "What an account holds and what it cost (average cost input)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "book id",
                        "name": "bookID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "account id",
                        "name": "accountID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD, default today",
                        "name": "as_of",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "transaction id to leave out (the one being edited)",
                        "name": "exclude",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.CostBasisDTO"
                        }
                    }
                }
            }
        },
...
        "/api/books/{bookID}/commodities": {
            "post": {
                "description": "Commodities are global like rates; the book only authorises the caller.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reference"
                ],
                "summary": "Add a security or a points programme (editor of this book)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "book id",
                        "name": "bookID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "commodity",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.createCommodityRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/routes.CommodityDTO"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
...
        "/api/me/account": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "Change the username (needs the current password); email changes go through /api/me/email",
                "parameters": [
                    {
                        "description": "identity",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.identityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.UserDTO"
                        }
                    },
                    "409": {
                        "description": "username taken",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
...
        "/api/me/password": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "Change password and sign out every other session",
                "parameters": [
                    {
                        "description": "passwords",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.passwordRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
...
        "/api/me/settings": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "Replace the user's preferences",
                "parameters": [
                    {
                        "description": "all preference fields",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.settingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.UserDTO"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "db.CommodityKind": {
            "type": "string",
            "enum": [
                "currency",
                "security",
                "points"
            ],
            "x-enum-varnames": [
                "CommodityKindCurrency",
                "CommodityKindSecurity",
                "CommodityKindPoints"
            ]
        },
...
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorDetail"
                }
            }
        },
        "response.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "routes.CommodityDTO": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "contract_size": {
                    "type": "string"
                },
                "decimals": {
                    "type": "integer"
                },
                "exchange_mic": {
                    "type": "string"
                },
                "kind": {
                    "$ref": "#/definitions/db.CommodityKind"
                },
                "name": {
                    "type": "string"
                },
                "quote_currency": {
                    "type": "string"
                }
            }
        },
        "routes.CostBasisDTO": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                },
                "unit_cost": {
                    "type": "string"
                }
            }
        },
...
        "routes.UserDTO": {
            "type": "object",
            "properties": {
                "date_format": {
                    "type": "string"
                },
                "default_book_id": {
                    "type": "integer"
                },
                "display_currency": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "initialized": {
                    "description": "False until the first-login wizard is done; most of the API answers\n403 onboarding_required meanwhile.",
                    "type": "boolean"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "language": {
                    "type": "string"
                },
                "password_must_change": {
                    "type": "boolean"
                },
                "pending_email": {
                    "description": "An address a verification code was sent to and not yet confirmed\n(only on GET /api/me).",
                    "type": "string"
                },
                "theme": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "routes.createCommodityRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "contract_size": {
                    "type": "string"
                },
                "decimals": {
                    "type": "integer"
                },
                "exchange_mic": {
                    "type": "string"
                },
                "kind": {
                    "$ref": "#/definitions/db.CommodityKind"
                },
                "name": {
                    "type": "string"
                },
                "quote_currency": {
                    "type": "string"
                }
            }
        },
        "routes.identityRequest": {
            "type": "object",
            "properties": {
                "current_password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "routes.loginRequest": {
            "type": "object",
            "properties": {
                "client": {
                    "description": "\"api\" returns a bearer token in the body instead of setting a cookie.",
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "routes.loginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/routes.UserDTO"
                }
            }
        },
...
        "routes.passwordRequest": {
            "type": "object",
            "properties": {
                "current_password": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                }
            }
        },
...
        "routes.settingsRequest": {
            "type": "object",
            "properties": {
                "date_format": {
                    "type": "string"
                },
                "default_book_id": {
                    "type": "integer"
                },
                "display_currency": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "theme": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
...
    }
}